+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Adds input parameters validation #168
 - Expands code coverage to 88% #147
 - Removes json assumption for request body to support multiple formats #170
+- Escapes language reserved keywords #184
 
 ## [0.0.4] - 2021-04-28